Add Get-InfobloxDNSAuthZones cmdlet
Introduced a PowerShell script to retrieve authorized DNS zones from Infoblox, enabling users to query zones based on FQDN and view parameters. This function checks for an active Infoblox server connection before proceeding with the query, displaying a warning if not connected. It supports lowercasing of FQDN and view values to ensure consistent query formatting. Additionally, it retrieves an exhaustive list of DNS zone attributes, enhancing the granularity of information available to administrators.